requirements.txt
python-dotenv
azure-identity
azure-ai-agents
fastapi
uvicorn
requests
beautifulsoup4
pydantic